training_scripts/requirements-training-rfdetr.txt
absl-py==2.4.0
accelerate==1.12.0
albucore==0.0.23
albumentations==1.4.24
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
bitsandbytes @ file:///home/thora/Downloads/bitsandbytes-0.48.0-cp312-cp312-linux_aarch64.whl#sha256=8d3a7ce3521d7acf9d16ddff1e7fa77e692c89a7d9e0cb924d4fb57831a184c7
blinker==1.9.0
cachetools==7.0.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
clearml==2.1.3
click==8.3.1
cloudpickle==3.1.2
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
databricks-sdk==0.94.0
defusedxml==0.7.1
docker==7.1.0
einops==0.8.2
fastapi==0.134.0
filelock==3.24.3
filetype==1.2.0
flash_attn @ file:///home/thora/Downloads/flash_attn-2.8.4-cp312-cp312-linux_aarch64.whl#sha256=d98079d4461bfefc31536b2129050377f0d9b21a80335c05cefd9d4a9ada60d1
Flask==3.1.3
flask-cors==6.0.2
fonttools==4.61.1
fsspec==2026.2.0
furl==2.1.4
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.48.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenlet==3.3.2
grpcio==1.78.0
gunicorn==25.1.0
h11==0.16.0
hf-xet==1.3.2
huey==2.6.0
huggingface_hub==0.36.2
idna==3.7
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
Mako==1.3.10
Markdown==3.10.2
MarkupSafe==3.0.3
matplotlib==3.10.8
mlflow==3.10.0
mlflow-skinny==3.10.0
mlflow-tracing==3.10.0
mpmath==1.3.0
networkx==3.6.1
numpy==2.2.6
opencv-contrib-python-rolling @ file:///home/thora/Downloads/opencv_contrib_python_rolling-4.13.0-cp312-cp312-linux_aarch64.whl#sha256=6e77b9ad7aeba994db0b443c047a4729c379a21617f64497c0d22f992d9b7be2
opencv-python==4.13.0.92
opencv-python-headless==4.10.0.84
opentelemetry-api==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orderedmultidict==1.0.2
packaging==26.0
pandas==2.3.3
pathlib2==2.3.7.post1
peft==0.18.1
pi_heif==1.3.0
pillow==12.1.1
pillow-avif-plugin==1.5.5
platformdirs==4.9.2
prettytable==3.17.0
protobuf==6.33.5
psutil==7.2.2
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycocotools==2.0.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyDeprecate==0.5.0
PyJWT==2.10.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rf100vl==1.1.0
rfdetr==1.5.1
roboflow==1.2.15
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentry-sdk==2.53.0
setuptools==82.0.0
simsimd==6.5.13
six==1.17.0
skops==0.13.0
smmap==5.0.2
SQLAlchemy==2.0.47
sqlparse==0.5.5
starlette==0.52.1
stringzilla==4.6.0
supervision==0.27.0.post1
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.22.2
torch @ file:///home/thora/Downloads/torch-2.9.1-cp312-cp312-linux_aarch64.whl#sha256=e6521b662893847896ece6740a6f3bd996c8a067d7b3334aadd6e3dcef47fd64
torchaudio @ file:///home/thora/Downloads/torchaudio-2.9.1-cp312-cp312-linux_aarch64.whl#sha256=da9b4df0a43fcda72d45ec546146a5340de4d0a0aad4c13a2fe4ef0809903907
torchvision @ file:///home/thora/Downloads/torchvision-0.24.1-cp312-cp312-manylinux_2_28_aarch64.whl#sha256=7fb7590c737ebe3e1c077ad60c0e5e2e56bb26e7bccc3b9d04dbfc34fd09f050
tqdm==4.67.3
transformers==4.57.6
triton @ file:///home/thora/Downloads/triton-3.5.0-cp312-cp312-linux_aarch64.whl#sha256=96829c4f038c9bbd90e240c877c124477625775236b796ba508b8bec9dde2902
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
wandb==0.25.0
wcwidth==0.6.0
Werkzeug==3.1.6
xformers @ file:///home/thora/Downloads/xformers-0.0.33.post1-cp39-abi3-linux_aarch64.whl#sha256=46896491592d19ee98203c1fc3734ec5a895bcd5a352edbd2a09083c1311398c
zipp==3.23.0
